[core] Moved method not allowed logic to an util class
